requirements.txt
feedparser>=6.0.11
beautifulsoup4>=4.12.0
notion-client>=2.0.0
python-dotenv>=1.0.0
